+++ bokeh/application/handlers/tests/test_notebook.py
@@ -24,7 +24,7 @@ import nbformat
 
 # Bokeh imports
 from bokeh.document import Document
-from bokeh.testing.util import with_temporary_file
+from bokeh.testing.filesystem import with_temporary_file
 
 # Module under test
 import bokeh.application.handlers.notebook as bahn
